+++ src/main/java/com/dailyon/snsservice/dto/request/comment/CreateReplyCommentRequest.java
@@ -1,4 +1,4 @@
-package com.dailyon.snsservice.dto.request.post;
+package com.dailyon.snsservice.dto.request.comment;
 
 
 import javax.validation.constraints.NotBlank;
